# openapi/openapi.yaml
openapi: 3.0.0
info:
  title: API
  version: 1.0.0
paths:
  /tasks:
    get:
      summary: Get all tasks
      tags:
        - tasks
      responses:
        '200':
          description: A list of tasks
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Task'
    post:
      summary: Create a new task
      tags:
        - tasks
      requestBody:
        description: JSON body to create a task
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateTaskRequest'
      responses:
        '201':
          description: The created task
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Task'
  /tasks/{id}:
    patch:
      summary: Update a task
      tags: 
        - tasks
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
            format: uint
      requestBody:
        description: JSON body to update a task
        required: true
        content: 
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateTaskRequest'
      responses:
        '200':
          description: Updated task
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Task'
    delete:
      summary: Delete a task by ID
      tags:
        - tasks
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: uint
      responses:
        '204':
          description: Task deleted successfully
        '404':
          description: Task not found
...
components:
  schemas:
    Task:
      type: object
      properties:
        id:
          type: integer
          format: uint
        task:
          type: string
        is_done:
          type: boolean
        user_id:
          type: integer
          format: uint
    CreateTaskRequest:
      type: object
      required:
        - task
        - user_id
      properties:
        task:
          type: string
        is_done:
          type: boolean
          nullable: true
        user_id:
          type: integer
          format: uint
    UpdateTaskRequest:
      type: object
      properties:
        task:
          type: string
          nullable: true
        is_done:
          type: boolean
          nullable: true
        user_id:
          type: integer
          format: uint
          nullable: true
...